_scripts: do not remove hugo target in serve mode
In https://cuelang.org/cl/1169000 we introduced a hack to remove
hugo/content when running the preprocessor with the execute command.

That had the side effect of also removing hugo/content when the serve
script is running (because it runs hugo execute --serve).

This, however, triggers an issue with ref and relref shortcodes.

The reasons it triggers an issue is as follows:

* hugo/content is empty post removal.
* preprocessor execute --serve runs which starts hugo.
* hugo sees an empty hugo/content to start with.
* preprocessor concurrently processes pages, which means contents of
  hugo/content get written in a random order.
* hugo watch therefore sees varying states of completeness of
  hugo/content, which can mean it reads the contents at a time when the
  target of a ref or relref do not yet exist.

We should ultimately fix this by using preprocessor directives instead
of ref and relref shortcodes, but for now we simply build on the earlier
hack to avoid removing hugo/content if we are in serve mode.

Preprocessor-No-Write-Cache: true
Reviewed-on: https://review.gerrithub.io/c/cue-lang/cuelang.org/+/1169582
TryBot-Result: CUEcueckoo <cueckoo@cuelang.org>

+++ _scripts/runPreprocessor.bash
@@ -28,9 +28,11 @@ GOBIN=$PWD/.gobin go install -trimpath -buildvcs=false ./internal/cmd/preprocess
 
 # Remove preprocessor target directory. This ensures we don't leave any stale
 # files lying around. Slight hack: only do so if we are running the execute command.
+# Note this ties this script to _scripts/serve.bash so if making changes to order
+# or args in that script, make changes to the check here.
 #
 # TODO: make the prepreprocess smart enough to do this itself.
-if [[ "${1:-}" == "execute" ]]; then
+if [[ "${1:-}" == "execute" && "${2:-}" != "--serve" ]]; then
 	rm -rf hugo/content
 fi
